btf: move btfHeader to btf_types.go
Move btfHeader so that all btf* types live in the same file. This
reduces the size of btf.go as well.

No other changes intended.
